Adds authentication activity, fragment, controller and services
Added rough implementation of an authentication activity. The fragment
displays a basic ui containing two text fields and a submit button to
the user. The user can then insert his credentials and hit the submit
button to start the authentication process.
The result of the authentication process is not yet evaluated.